move js back into vue component
