+++ pages/narratives/_book/_chapter/index.vue
@@ -4,4 +4,27 @@
     <card-list v-bind:cards="pages" />
   </div>
 </template>
-<script src="./index-component.js"></script>
+<script>
+import CardList from '~/components/card-list/CardList'
+import loadData from '~/lib/load-data'
+import events from '~/components/events/events'
+
+export default {
+  async asyncData (context) {
+    const { pages, title } = await loadData(context, context.params)
+    return { pages, title }
+  },
+  mounted () {
+    const marker = {
+      slug: this.slug,
+      location: this.location,
+      path: this.path
+    }
+    this.$events.$emit(events.activeMarkerChanged, marker)
+    this.$events.$emit(events.markersChanged, undefined)
+  },
+  components: {
+    CardList
+  }
+}
+</script>
